Add etag support

+++ src/Microsoft.Restier.Publishers.OData/RestierController.cs
@@ -1,7 +1,9 @@
 // Copyright (c) Microsoft Corporation.  All rights reserved.
 // Licensed under the MIT License.  See License.txt in the project root for license information.
+extern alias Net;
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -25,10 +27,12 @@ using Microsoft.Restier.Core.Submit;
 using Microsoft.Restier.Publishers.OData.Batch;
 using Microsoft.Restier.Publishers.OData.Filters;
 using Microsoft.Restier.Publishers.OData.Formatter;
-using Microsoft.Restier.Publishers.OData.Formatter.Deserialization;
 using Microsoft.Restier.Publishers.OData.Properties;
 using Microsoft.Restier.Publishers.OData.Query;
 using Microsoft.Restier.Publishers.OData.Results;
+
+// This is a must for creating response with correct extension method
+using Net::System.Net.Http;
 
 namespace Microsoft.Restier.Publishers.OData
 {
@@ -37,10 +41,10 @@ namespace Microsoft.Restier.Publishers.OData
     /// </summary>
     [RestierFormatting]
     [RestierExceptionFilter]
-    public sealed class RestierController : ODataController
+    public class RestierController : ODataController
     {
-        private const string ETagGetterKey = "ETagGetter";
-        private const string ETagHeaderKey = "@etag";
+        private const string IfMatchKey = "@IfMatchKey";
+        private const string IfNoneMatchKey = "@IfNoneMatchKey";
 
         private ApiBase api;
         private bool shouldReturnCount;
@@ -81,6 +85,10 @@ namespace Microsoft.Restier.Publishers.OData
 
             // Get queryable path builder to builder
             IQueryable queryable = this.GetQuery(path);
+            ETag etag;
+
+            // TODO This flag can be removed when Etag isIfNoneMatch is changed to public.
+            bool isIfNoneMatch;
 
             // TODO #365 Do not support additional path segment after function call now
             if (lastSegment.SegmentKind == ODataSegmentKinds.UnboundFunction)
@@ -89,7 +97,7 @@ namespace Microsoft.Restier.Publishers.OData
                 Func<string, object> getParaValueFunc = p => unboundSegment.GetParameterValue(p);
                 result = await ExecuteOperationAsync(
                     getParaValueFunc, unboundSegment.FunctionName, true, null, cancellationToken);
-                result = ApplyQueryOptions(result, path, true);
+                result = ApplyQueryOptions(result, path, true, out isIfNoneMatch, out etag);
             }
             else
             {
@@ -110,16 +118,16 @@ namespace Microsoft.Restier.Publishers.OData
                     result = await ExecuteOperationAsync(
                         getParaValueFunc, boundSeg.Function.Name, true, result, cancellationToken);
 
-                    result = ApplyQueryOptions(result, path, true);
+                    result = ApplyQueryOptions(result, path, true, out isIfNoneMatch, out etag);
                 }
                 else
                 {
-                    queryable = ApplyQueryOptions(queryable, path, false);
+                    queryable = ApplyQueryOptions(queryable, path, false, out isIfNoneMatch, out etag);
                     result = await ExecuteQuery(queryable, cancellationToken);
                 }
             }
 
-            return this.CreateQueryResponse(result, path.EdmType);
+            return this.CreateQueryResponse(result, path.EdmType, isIfNoneMatch, etag);
         }
 
         /// <summary>
@@ -155,6 +163,7 @@ namespace Microsoft.Restier.Publishers.OData
                 entitySet.Name,
                 expectedEntityType.GetClrType(Api),
                 actualEntityType.GetClrType(Api),
+                ChangeSetItemAction.Insert,
                 null,
                 null,
                 edmEntityObject.CreatePropertyDictionary());
@@ -223,12 +232,19 @@ namespace Microsoft.Restier.Publishers.OData
                 throw new NotImplementedException(Resources.DeleteOnlySupportedOnEntitySet);
             }
 
+            var propertiesInEtag = await this.GetOriginalValues(entitySet);
+            if (propertiesInEtag == null)
+            {
+                return this.StatusCode((HttpStatusCode)428);
+            }
+
             DataModificationItem deleteItem = new DataModificationItem(
                 entitySet.Name,
                 path.EdmType.GetClrType(Api),
                 null,
+                ChangeSetItemAction.Remove,
                 RestierQueryBuilder.GetPathKeyValues(path),
-                this.GetOriginalValues(),
+                propertiesInEtag,
                 null);
 
             RestierChangeSetProperty changeSetProperty = this.Request.GetChangeSet();
@@ -298,6 +314,8 @@ namespace Microsoft.Restier.Publishers.OData
                 if (lastSegment.SegmentKind == ODataSegmentKinds.Action)
                 {
                     var queryResult = await ExecuteQuery(queryable, cancellationToken);
+
+                    // TODO GitHubIssue#114, need etag check here for single bound entity
                     var boundSeg = (BoundActionPathSegment)lastSegment;
                     result = await ExecuteOperationAsync(
                         getParaValueFunc, boundSeg.Action.Name, false, queryResult, cancellationToken);
@@ -310,7 +328,7 @@ namespace Microsoft.Restier.Publishers.OData
                 return this.Request.CreateResponse(HttpStatusCode.NoContent);
             }
 
-            return this.CreateQueryResponse(result, path.EdmType);
+            return this.CreateQueryResponse(result, path.EdmType, false, null);
         }
 
         /// <summary>
@@ -366,6 +384,12 @@ namespace Microsoft.Restier.Publishers.OData
                 throw new NotImplementedException(Resources.UpdateOnlySupportedOnEntitySet);
             }
 
+            var propertiesInEtag = await this.GetOriginalValues(entitySet);
+            if (propertiesInEtag == null)
+            {
+                return this.StatusCode((HttpStatusCode)428);
+            }
+
             // In case of type inheritance, the actual type will be different from entity type
             // This is only needed for put case, and does not for patch case
             var expectedEntityType = path.EdmType;
@@ -380,8 +404,9 @@ namespace Microsoft.Restier.Publishers.OData
                 entitySet.Name,
                 expectedEntityType.GetClrType(Api),
                 actualEntityType.GetClrType(Api),
+                ChangeSetItemAction.Update,
                 RestierQueryBuilder.GetPathKeyValues(path),
-                this.GetOriginalValues(),
+                propertiesInEtag,
                 edmEntityObject.CreatePropertyDictionary());
             updateItem.IsFullReplaceUpdateRequest = isFullReplaceUpdate;
 
@@ -403,7 +428,8 @@ namespace Microsoft.Restier.Publishers.OData
             return this.CreateUpdatedODataResult(updateItem.Entity);
         }
 
-        private HttpResponseMessage CreateQueryResponse(IQueryable query, IEdmType edmType)
+        private HttpResponseMessage CreateQueryResponse(
+            IQueryable query, IEdmType edmType, bool isIfNoneMatch, ETag etag)
         {
             IEdmTypeReference typeReference = GetTypeReference(edmType);
 
@@ -475,8 +501,8 @@ namespace Microsoft.Restier.Publishers.OData
                     HttpStatusCode.OK, new EntityCollectionResult(query, typeReference, this.Api.Context));
             }
 
-            var entityResult = new EntityResult(query, typeReference, this.Api.Context);
-            if (entityResult.Result == null)
+            var entityResult = query.SingleOrDefault();
+            if (entityResult == null)
             {
                 // TODO GitHubIssue#288: 204 expected when requesting single nav property which has null value
                 // ~/People(nonexistkey) and ~/People(nonexistkey)/BestFriend, expected 404
@@ -487,8 +513,36 @@ namespace Microsoft.Restier.Publishers.OData
                         Resources.ResourceNotFound));
             }
 
-            // TODO GitHubIssue#43 : support non-Entity ($select/$value) queries
-            return this.Request.CreateResponse(HttpStatusCode.OK, entityResult);
+            // Check the ETag here
+            if (etag != null)
+            {
+                // request with If-Match header, if match, then should return whole content
+                // request with If-Match header, if not match, then should return 412
+                // request with If-None-Match header, if match, then should return 304
+                // request with If-None-Match header, if not match, then should return whole content
+                etag.EntityType = query.ElementType;
+                query = etag.ApplyTo(query);
+                entityResult = query.SingleOrDefault();
+                if (entityResult == null && !isIfNoneMatch)
+                {
+                    return this.Request.CreateResponse(HttpStatusCode.PreconditionFailed);
+                }
+                else if (entityResult == null)
+                {
+                    return this.Request.CreateResponse(HttpStatusCode.NotModified);
+                }
+            }
+
+            // Using reflection to create response for single entity so passed in parameter is not object type,
+            // but will be type of real entity type, then EtagMessageHandler can be used to set ETAG header
+            // when response is single entity.
+            // There are three HttpRequestMessageExtensions class defined in different assembles
+            var genericMethod = typeof(System.Net.Http.HttpRequestMessageExtensions).GetMethods()
+                .Where(m => m.Name == "CreateResponse" && m.GetParameters().Length == 3);
+            var method = genericMethod.FirstOrDefault().MakeGenericMethod(query.ElementType);
+            response = method.Invoke(null, new object[] { this.Request, HttpStatusCode.OK, entityResult })
+                as HttpResponseMessage;
+            return response;
         }
 
         private IQueryable GetQuery(ODataPath path)
@@ -501,8 +555,13 @@ namespace Microsoft.Restier.Publishers.OData
             return queryable;
         }
 
-        private IQueryable ApplyQueryOptions(IQueryable queryable, ODataPath path, bool applyCount)
+        private IQueryable ApplyQueryOptions(
+            IQueryable queryable, ODataPath path, bool applyCount, out bool isIfNoneMatch, out ETag etag)
         {
+            // ETAG IsIfNoneMatch is changed to public access, this flag can be removed.
+            isIfNoneMatch = false;
+            etag = null;
+
             if (this.shouldWriteRawValue)
             {
                 // Query options don't apply to $value.
@@ -513,6 +572,17 @@ namespace Microsoft.Restier.Publishers.OData
             ODataQueryContext queryContext =
                 new ODataQueryContext(properties.Model, queryable.ElementType, path);
             ODataQueryOptions queryOptions = new ODataQueryOptions(queryContext, this.Request);
+
+            // Get etag for query request
+            if (queryOptions.IfMatch != null)
+            {
+                etag = queryOptions.IfMatch;
+            }
+            else if (queryOptions.IfNoneMatch != null)
+            {
+                isIfNoneMatch = true;
+                etag = queryOptions.IfNoneMatch;
+            }
 
             // TODO GitHubIssue#41 : Ensure stable ordering for query
             ODataQuerySettings settings = Api.Context.GetApiService<ODataQuerySettings>();
@@ -559,8 +629,6 @@ namespace Microsoft.Restier.Publishers.OData
             };
 
             QueryResult queryResult = await Api.QueryAsync(queryRequest, cancellationToken);
-
-            this.Request.Properties[ETagGetterKey] = this.Api.Context.GetProperty(ETagGetterKey);
             var result = queryResult.Results.AsQueryable();
             return result;
         }
@@ -596,7 +664,7 @@ namespace Microsoft.Restier.Publishers.OData
             return result;
         }
 
-        private IReadOnlyDictionary<string, object> GetOriginalValues()
+        private async Task<IReadOnlyDictionary<string, object>> GetOriginalValues(IEdmEntitySet entitySet)
         {
             Dictionary<string, object> originalValues = new Dictionary<string, object>();
 
@@ -606,7 +674,26 @@ namespace Microsoft.Restier.Publishers.OData
                 ETag etag = this.Request.GetETag(etagHeaderValue);
                 etag.ApplyTo(originalValues);
 
-                originalValues.Add(ETagHeaderKey, etagHeaderValue.Tag);
+                originalValues.Add(IfMatchKey, etagHeaderValue.Tag);
+                return originalValues;
+            }
+
+            etagHeaderValue = this.Request.Headers.IfNoneMatch.SingleOrDefault();
+            if (etagHeaderValue != null)
+            {
+                ETag etag = this.Request.GetETag(etagHeaderValue);
+                etag.ApplyTo(originalValues);
+
+                originalValues.Add(IfNoneMatchKey, etagHeaderValue.Tag);
+                return originalValues;
+            }
+
+            // return 428(Precondition Required) if entity requires concurrency check.
+            var model = await this.Api.Context.GetModelAsync();
+            bool needEtag = model.IsConcurrencyCheckEnabled(entitySet);
+            if (needEtag)
+            {
+                return null;
             }
 
             return originalValues;
